feat: add logging decorator to enhance method call tracing in job scrapers
